Added authorization

+++ routes/users.js
@@ -4,18 +4,7 @@ const catchAsync = require('../utils/catchAsync.js');
 const ExpressError = require('../utils/ExpressError.js');
 const User = require("../models/user.js");
 const {userSchema} = require('../schemas.js');
-const {isLoggedIn} = require('../middleware.js');
-
-const validateUser = (req, res, next) => {
-    const {error} = userSchema.validate(req.body);
-    if(error){
-        const msg = error.details.map(el=>el.message).join(',');
-        throw new ExpressError(msg, 400);
-    }
-    else{
-        next();
-    }
-}
+const {isLoggedIn, validateUser} = require('../middleware.js');
 
 // Users Index Route
 Router.get("/", catchAsync(async (req, res) => {
@@ -35,6 +24,7 @@ Router.post("/", isLoggedIn, validateUser, catchAsync(async (req, res, next) => 
     //     throw new ExpressError("Invalid image URL provided", 400);
     // }
     const newUser = new User(user);
+    newUser.author = req.user._id;
     await newUser.save();
     req.flash('success', "Successfully added a new user!!");
     res.redirect(`/users/${newUser._id}`);
@@ -43,7 +33,7 @@ Router.post("/", isLoggedIn, validateUser, catchAsync(async (req, res, next) => 
 
 // Show User Route
 Router.get("/:id", catchAsync(async (req, res) => {
-    const user = await User.findById(req.params.id).populate('reviews');
+    const user = await User.findById(req.params.id).populate({path: 'reviews', populate: {path: 'author'}}).populate('author');
     if (!user) {
         req.flash('error', "Cannot find the user!!");
         return res.redirect('/users');
@@ -52,21 +42,29 @@ Router.get("/:id", catchAsync(async (req, res) => {
 }));
 
 // Edit User Form Route
-Router.get("/:id/edit", catchAsync(async (req, res) => {
-    const user = await User.findById(req.params.id);
+Router.get("/:id/edit", isLoggedIn, catchAsync(async (req, res) => {
+    const { id } = req.params;
+    const user = await User.findById(id);
     if (!user) {
         req.flash('error', "Cannot find the user!!");
         return res.redirect('/users');
     }
+    if (!user.author || !user.author.equals(req.user._id)) {
+        req.flash('error', "You do not have the permission to do that!!");
+        return res.redirect(`/users/${id}`);
+    }
     res.render("users/edit", { user });
 }));
+
 
 // Update User
 Router.put("/:id", validateUser, catchAsync(async (req, res) => {
     const { id } = req.params;
     const { user } = req.body;
-    if (user.profileImageURL && !isValidUrl(user.profileImageURL)) {
-        throw new ExpressError("Invalid image URL provided", 400);
+    const user1 = await User.findById(id);
+    if (!user1.author || !user1.author.equals(req.user._id)) {
+        req.flash('error', "You do not have the permission to do that!!");
+        return res.redirect(`/users/${id}`);
     }
     await User.findByIdAndUpdate(id, { ...user });
     req.flash('success', "Successfully updated the user!!");
@@ -77,6 +75,11 @@ Router.put("/:id", validateUser, catchAsync(async (req, res) => {
 // Delete User
 Router.delete("/:id", catchAsync(async (req, res) => {
     const { id } = req.params;
+    const user1 = await User.findById(id);
+    if (!user1.author || !user1.author.equals(req.user._id)) {
+        req.flash('error', "You do not have the permission to do that!!");
+        return res.redirect(`/users/${id}`);
+    }
     await User.findByIdAndDelete(id);
     req.flash('success', "Aww, we deleted the user!!");
     res.redirect("/users");
